CHeaders, CommonCode: Add pointer read/write functions
I was seeing lots of cases where the last bytes of a buffer got cast to
an integer, or back again.  As the buffer was an array of UINTs, this
was fraught with compatibility problems.

To avoid problems, we'll use `sizeof(void*)` as our measure for
pointer-width, and use `memcpy` to copy the value in and out of the
buffer.  I've added the ability to grab other pointers at the end of the
buffer too, just in case that's needed.